+++ src/js/components/weather/user-input/temperature/TemperatureRange.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { Range } from "rc-slider";
 
-import { weatherActions, convenience } from '../WeatherActions';
-import { weatherStore } from '../WeatherStore';
+import { weatherActions, convenience } from 'js/components/weather/WeatherActions';
+import { weatherStore } from 'js/components/weather/WeatherStore';
 
 import 'rc-slider/assets/index.css'
 
@@ -19,7 +19,9 @@ class TemperatureRange extends Component {
     handleAfterChange(range) {
         convenience.isNotGrabbing();
         weatherActions.temperatureRange(range);
-        weatherActions.youShouldBike();
+        if (weatherStore.getState().weather) {
+            weatherActions.youShouldBike();
+        }
     }
 
     render() {
